\name{ambur.setup}
\alias{ambur.setup}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Install required R packages }
\description{
Installs required R packages for AMBUR
}
\usage{
ambur.setup(setup = 1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{setup }{A dummy variable set to 1. Do not change }
}
\details{
Installs required R packages to run AMBUR
}
\value{

Installation of R packages in library
}
\references{ }
\note{}
\seealso{  }

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ stat }
\keyword{ shapefile }% __ONLY ONE__ keyword per line