% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number}

\item{year}{Year that data were collected}
}
\value{
Visual map
}
\description{
Create map visualization of FARS data
}
\examples{
\dontrun{
fars_map_state(1, 2013)
}
}
