% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{FARS data import}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{File (CSV) name}
}
\value{
Data frame table
}
\description{
Reads in FARS data
}
\examples{
\dontrun{
fars_read("accident_2013.csv")
}

}
