% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read Multiple Files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Years that data were collected}
}
\value{
List of data frames (NULL if data do not exist)
}
\description{
Read multiple files and confirm that the years are valid
}
\examples{
\dontrun{
fars_read_years(2013:2015)
}

}
