% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Years that data were collected}
}
\value{
Summary
}
\description{
Summarize yearly data by month
}
\examples{
\dontrun{
fars_summarize_years(2013:2015)
}

}
