% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun5.R
\name{Fun5}
\alias{Fun5}
\title{A function for testing whether the sum of two numbers is greater than 5}
\usage{
Fun5(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{another number}
}
\value{
a number
}
\description{
A really good test function.
Perhaps the best function ever!

A work of pure genius.
}
