% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_from}
\alias{import_from}
\title{Import functions from package}
\usage{
import_from(.from, ...)
}
\arguments{
\item{.from}{The package from which to import, or the path to a stand-alone
.R file.}

\item{...}{Names or name-value pairs specifying objects to import.
If arguments are named, then the imported object will have this new name.}
}
\description{
Alternative to library.
}
\examples{
# import_from(magrittr, "\%>\%")
# import_from(checkmate, assert_logical)
}
\seealso{
import_pkg
}
