% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_pkg}
\alias{import_pkg}
\title{Import package}
\usage{
import_pkg(.pkg, .load = FALSE, .install = FALSE)
}
\arguments{
\item{.pkg}{Character scalar. Package name.}

\item{.load}{Logical scalar. Load package?}

\item{.install}{Logical scalar. Install package if not already installed?}
}
\description{
Alternative to library.
}
\examples{
# import_pkg("lubridate")
# import_pkg("lubridate", .load = TRUE)
}
