% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages_if_needed_and_load.R
\name{install_packages_if_needed_and_load}
\alias{install_packages_if_needed_and_load}
\title{Install package if needed and load}
\usage{
install_packages_if_needed_and_load(pkgs)
}
\arguments{
\item{pkgs}{character vector of the names of packages whose current versions should be downloaded from the repositories.}
}
\value{
nothing
}
\description{
Install all the non installed packages among a vector of package names
and load them all.
}
\examples{
install_packages_if_needed_and_load(c("rlang", "lobstr"))
}
