% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate.R
\name{investigate}
\alias{investigate}
\title{A function for plotting a histogram and finding numeric summaries of data}
\usage{
investigate(values)
}
\arguments{
\item{values}{A vector of values}
}
\value{
A tibble of numeric summaries and a histogram plot
}
\description{
A really good multi purpose function.
Perhaps the best function ever!

A work of pure genius.
}
\examples{
investigate(movies$duration)
investigate(movies$rating)
}
