% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Year that data were collected}
}
\value{
Filename
}
\description{
Create filename for FARS data
}
\examples{
\dontrun{
make_filename(2017)
}

}
