% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\alias{times}
\title{A function for multiplying}
\usage{
multiply(x, y)

times(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{another number}
}
\value{
a number
}
\description{
A really good multiplying function.
Perhaps the best function ever!

A work of pure genius.
}
\examples{
multiply(5, 10)
## Can also use negative numbers
multiply(-5, 10)
times(4, 3)
times(-4, 2)
}
