% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-package.R
\docType{package}
\name{pkg-package}
\alias{pkg}
\alias{pkg-package}
\title{pkg: Package Objects}
\description{
What if we could encapsulate a package in an R6 object ?
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moodymudskipper/pkg}
  \item Report bugs at \url{https://github.com/moodymudskipper/pkg/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
