% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_pkg_example.R
\docType{data}
\name{pkg_example}
\alias{pkg_example}
\title{pkg_example}
\format{
A data frame with 8 rows and 3 variables:
\describe{
\item{ name }{  character }
\item{ sex }{  character }
\item{ gender }{  character }
}
}
\source{
dplyr Package
}
\usage{
pkg_example
}
\description{
A sample of white skin people from starwars.
}
\keyword{datasets}
