% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{pkg_info}
\alias{pkg_info}
\title{Get information about a package}
\usage{
pkg_info(pkg)
}
\arguments{
\item{pkg}{The package of interest.}
}
\value{
A named vector. It has the same entries as the output of
utils::installed.packages, i.e., "Package", "LibPath",
"Version", "Priority", "Depends", "Imports", "LinkingTo",
"Suggests", "Enhances", "OS_type", "License" and "Built".
}
\description{
Get information about a package such as its version, license, etc.
}
\author{
Ruiyang Wu
}
