% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{pkg_install}
\alias{pkg_install}
\title{Install packages}
\usage{
pkg_install(pkgs, ...)
}
\arguments{
\item{pkgs}{A vector of packages to be installed}
}
\value{
A vector containing packages user requests to install
(invisible)
}
\description{
Install packages
}
\author{
Ruiyang Wu
}
