% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{pkg_leaves}
\alias{pkg_leaves}
\title{List all leaf packages}
\usage{
pkg_leaves()
}
\value{
A vector consisting of all leaf packages installed.
}
\description{
List the names of all leaf packages installed.
}
\author{
Ruiyang Wu
}
