% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{pkg_list}
\alias{pkg_list}
\title{List installed packages}
\usage{
pkg_list(user_installed = TRUE, exclude_recommended = TRUE)
}
\arguments{
\item{user_installed}{A Boolean variable indicating whether only
user-installed packages are listed. If FALSE, list all installed
packages.}

\item{include_recommended}{A Boolean variable indicating whether
"recommended" packages should be excluded from the list.}
}
\value{
A character vector of the requested packages.
}
\description{
List installed packages.
}
\author{
Ruiyang Wu
}
