% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{pkg_list_user}
\alias{pkg_list_user}
\title{List all user-installed packages}
\usage{
pkg_list_user()
}
\value{
A vector consisting of all user-installed packages.
}
\description{
List all user-installed packages.
}
\author{
Ruiyang Wu
}
