% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{pkg_purge}
\alias{pkg_purge}
\title{Purge a package}
\usage{
pkg_purge(pkg)
}
\arguments{
\item{pkg}{The package to be purged}
}
\value{
The purged package name (invisible)
}
\description{
Purge an installed package and its dependencies
}
\author{
Ruiyang Wu
}
