% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_test_internal_functions.R
\name{pkg_test_internal_functions}
\alias{pkg_test_internal_functions}
\title{Test d'usage des fonctions internes}
\usage{
pkg_test_internal_functions(pkgs)
}
\arguments{
\item{pkgs}{character vector of the names of packages whose current versions should be downloaded from the repositories.}
}
\value{
invisible NULL.
}
\description{
Teste l'usage des fonctions internes définies dans le chunk de définition de la fonction install_packages_if_needed and load
}
\examples{
pkg_test_internal_functions("lobstr")
}
