% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{pkg_uninstall}
\alias{pkg_uninstall}
\title{Uninstall packages}
\usage{
pkg_uninstall(pkgs, ...)
}
\arguments{
\item{pkgs}{A vector of packages to be uninstalled.}
}
\value{
The (invisible) same vector.
}
\description{
Uninstall packages
}
\author{
Ruiyang Wu
}
