% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.R
\name{subtract}
\alias{subtract}
\title{A function for subtracting}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{another number}
}
\value{
a number
}
\description{
A really good subtracting function.
Perhaps the best function ever!

A work of pure genius.
}
\examples{
subtract(5, 10)
## Can also use negative numbers
subtract(-5, 10)
}
