% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{pkg_list}
\alias{pkg_list}
\title{List all packages}
\usage{
pkg_list()
}
\value{
A vector consisting of all packages installed.
}
\description{
List the names of all packages installed.
}
\author{
Ruiyang Wu
}
